+++ kermit/k95/ckoco3.c
@@ -188,6 +188,8 @@ extern int tt_status[VNUM];           /* Terminal status line displayed */
 extern int tt_status_usr[VNUM];
 extern int tt_modechg;          /* Terminal Video-Change (80 or 132 cols) */
 extern int tt_senddata;         /* May data be sent to the host */
+extern int tt_clipboard_read = CLIPBOARD_DENY_NOTIFY,
+           tt_clipboard_write = CLIPBOARD_DENY_NOTIFY; /* OSC-52 */
 extern int tt_hidattr;          /* Attributes do not occupy a space */
 #ifdef PCTERM
 extern int tt_pcterm;
@@ -12422,8 +12424,267 @@ doosc( void ) {
 	case 46: /* xterm - change log file */
 	case 50: /* xterm - set font */
 	case 51: /* xterm - reserved for emacs shell */
-	case 52: /* xterm - manipulate selection data */
-             /* teraterm only implements the  clipboard feature */
+	case 52: { /* xterm - manipulate selection data */
+        /* Format is OSC 52 ; Pc ; Pd ST
+           Where Pc is some combination of zero or more of:
+                c, p, q, s, 0, 1, 2, 3, 4, 5, 6, 7
+           This specifies which of xterms various buffers we're manipulating.
+           Windows only has one (the clipboard), and its not worth emulating
+           the rest unless someone can point at an application that uses them.
+           So we just ignore this first parameter */
+
+        /* Don't use any of the code here as a good example for how to properly
+            do character set/unicode conversion in Kermit 95. I just copied
+            how it *appears* to be done elsewhere - I don't know if any of this
+            is truly correct -- DG */
+
+#ifdef NT
+        int use_unicode = (ck_isunicode() && !isWin95());
+#endif /* NT */
+
+        achar = (apcnext<=apclength)?apcbuf[apcnext++]:0;
+        while (strchr("cps01234567", achar)) {
+            achar = (apcnext<=apclength)?apcbuf[apcnext++]:0;
+        }
+
+        if (achar == ';') {
+            /* Pd is either '?' to query the current clipboard contents, or
+             * it contains base64 encoded data to set the clipboard to.
+             */
+            achar = (apcnext<=apclength)?apcbuf[apcnext++]:0;
+
+            if (achar == '?' && apcnext == apclength) {
+
+                /* Its a query */
+                if (tt_clipboard_read >= CLIPBOARD_ALLOW) {
+                    char* clipboardData = 0;
+                    int clipboardDataLen = 0;
+
+#ifdef KUI
+#ifdef CK_SHELL_NOTIFY
+                    if (tt_clipboard_read == CLIPBOARD_ALLOW_NOTIFY) {
+                        KuiShowNotification(
+                            KUI_NOTIF_I_INFO,
+                            "Clipboard Read",
+                            "Clipboard read by remote host allowed");
+                    }
+#endif /* CK_SHELL_NOTIFY */
+#endif /* KUI */
+
+                    /* Get data from the clipboard. On Windows NT, this will
+                     * be Unicode text and will need converting to the remote
+                     * character set. On Windows 9x and OS/2, we just send the
+                     * clipboard text as-is after base64-encoding*/
+#ifdef NT
+                    if (use_unicode) {
+                        USHORT * pUClipbrdData = GetUnicodeClipboardContent();
+
+                        /* Convert from UCS-2 to whatever the remote charset is */
+                        if ( pUClipbrdData ) {
+                            int i=0, j=0, len = 0, nbytes;
+                            unsigned char * bytes;
+
+                            len = wcslen(pUClipbrdData);
+                            debug(F111,"Clipboard","pUClipbrdData length",len);
+
+                            nbytes = utorxlat(pUClipbrdData[0], &bytes);
+                            if ( nbytes > 0 )
+                                clipboardDataLen = nbytes;
+                            for ( i=1; i<len; i++ ) {
+                                if ( pUClipbrdData[i-1] != CK_CR || pUClipbrdData[i] != LF ) {
+                                    nbytes = utorxlat(pUClipbrdData[i], &bytes);
+                                    if ( nbytes > 0 )
+                                        clipboardDataLen += nbytes;
+                                }
+                            }
+
+                            clipboardData = (unsigned char *) malloc(clipboardDataLen+1);
+                            memset(clipboardData, 0, clipboardDataLen+1);
+
+                            nbytes = utorxlat(pUClipbrdData[0], &bytes);
+                            while ( nbytes-- > 0 )
+                                clipboardData[j++] = *bytes++;
+                            for ( i=1; i<len; i++ ) {
+                                if ( pUClipbrdData[i-1] != CK_CR || pUClipbrdData[i] != LF ) {
+                                    nbytes = utorxlat(pUClipbrdData[i], &bytes);
+                                    while ( nbytes-- > 0 )
+                                        clipboardData[j++] = *bytes++;
+                                }
+                            }
+
+                            free( pUClipbrdData ) ;
+                        }
+                    } else {
+#endif /* NT */
+                        clipboardData = GetClipboardContent();
+                        clipboardDataLen = strlen(clipboardData);
+#ifdef NT
+                    }
+#endif /* NT */
+
+                    if (clipboardData) {
+                        /* Allocate memory for the maximum length the base64
+                         * encoded data could be */
+                        int rc;
+                        int encodedLen = 1 + ceil(clipboardDataLen/3.0)*4;
+                        char* encodedData = malloc(encodedLen);
+                        memset(encodedData, 0, encodedLen);
+
+                        /* Encode it... */
+                        if ((rc = b8tob64(clipboardData, -1, encodedData, encodedLen)) >= 0) {
+                            /* Send it in the form: OSC 52 ; data ST */
+                            sendchars("\033]52;", 5);
+                            sendchars(encodedData, rc);
+                            sendchars("\033\\", 2);
+                        } else {
+                            debug(F111, "base64 encode of clipboard data failed", "rc", rc);
+                        }
+
+                        free(clipboardData);
+                        free(encodedData);
+                    }
+                }
+#ifdef KUI
+#ifdef CK_SHELL_NOTIFY
+                else if (tt_clipboard_read == CLIPBOARD_DENY_NOTIFY) {
+                    KuiShowNotification(
+                        KUI_NOTIF_I_WARN,
+                        "Clipboard Read",
+                        "Clipboard read by remote host denied. You can enable "
+                        "clipboard access (or disable this "
+                        "notification) with the SET TERM CLIPBOARD-ACCESS "
+                        "command");
+                }
+#endif /* CK_SHELL_NOTIFY */
+#endif /* KUI */
+            } else {
+                /* Its a clipboard write */
+                if (tt_clipboard_write >= CLIPBOARD_ALLOW) {
+                    /* +1 for null termination, +1 because apcnext is already
+                     * pointing one character in */
+                    int cliplen = (apclength - apcnext) + 2;
+                    char* encoded = malloc(cliplen);
+                    char* decoded = malloc(cliplen);
+                    int rc;
+
+#ifdef KUI
+#ifdef CK_SHELL_NOTIFY
+                    if (tt_clipboard_write == CLIPBOARD_ALLOW_NOTIFY) {
+                        KuiShowNotification(
+                            KUI_NOTIF_I_INFO,
+                            "Clipboard Write",
+                            "Clipboard write by remote host allowed");
+                    }
+#endif /* CK_SHELL_NOTIFY */
+#endif /* KUI */
+
+                    strncpy(encoded, apcbuf+apcnext - 1, cliplen);
+
+                    /* Base64 decode the clipboard data and set it */
+                    rc = b64tob8(encoded, cliplen, decoded, cliplen);
+                    if (rc > 0) {
+                        /* Worst case: each UTF-8 character becomes one UCS-2
+                         * character. */
+                        USHORT* ucs2_string = malloc(sizeof(USHORT) * cliplen);
+                        memset(ucs2_string, 0, sizeof(USHORT) * cliplen);
+
+                        /* Ok, now we have a slight difficulty: the new data for
+                         * the clipboard has bypassed all normal characterset
+                         * translation/unicode conversion because it was hidden
+                         * away in base64 form. So now we've got to do all that
+                         * work here */
+
+                        /* The code here all works with UCS-2 because,
+                         * unfortunately, thats all Kermit 95 supports at
+                         * present. At some point we need a utf8-to-utf16
+                         * conversion function so we can support more than just
+                         * the basic multilingual plane, but thats a big job. */
+
+                        if (tt_utf8) {
+                            /* We're in UTF-8 mode - everything else the host is
+                             * sending us is assumed to be in UTF-8, so the new
+                             * text for the clipboard probably is too. So we
+                             * now need to convert it to UCS-2 to hand off to
+                             * Windows. */
+
+                            int utf8_idx, ucs2_idx = 0, rc;
+                            USHORT *us = NULL;
+
+                            for (utf8_idx = 0; decoded[utf8_idx] != '\0'; utf8_idx++) {
+                                rc = utf8_to_ucs2(decoded[utf8_idx], &us);
+                                if (rc == 0) {
+                                    /* UTF-8 sequence decoded, we have a UCS-2
+                                       character */
+                                    ucs2_string[ucs2_idx] = *us;
+                                    ucs2_idx++;
+                                } else if (rc < 0) {
+                                    /* Decoding failed. Output U+FFFD */
+                                    ucs2_string[ucs2_idx] = 0xfffd;
+                                    ucs2_idx++;
+                                }
+                                /* Else more UTF-8 bytes are needed to assemble
+                                 * the UCS-2 character. Continue. */
+                            }
+                        } else {
+                            int i;
+                            /* Remote is some 7-bit or 8-bit character set.
+                             * Windows wants UCS-2, so we need to convert it. */
+
+                            for (i = 0; decoded[i] != '\0'; i++) {
+                                if (decoded[i] >= 128)
+                                    ucs2_string[i] = (*xl_u[tcsr])(decoded[i]);
+                                else
+                                    ucs2_string[i] = decoded[i];
+                                    /* Some other code passes it through TX_IBMC0GRPH
+                                        if its not a control character we act on */
+                            }
+                        }
+
+#ifdef NT
+                        if (use_unicode) {
+                            rc = CopyToClipboard((BYTE*)ucs2_string,
+                                                 sizeof(USHORT) * cliplen);
+                        } else {
+#endif /* NT */
+                            /* We've now got the clipboard data as a UCS-2
+                             * string, but we're on Windows 9x or OS/2 (or NT
+                             * with Unicode support turned off), so we need to
+                             * now convert it back from UCS-2 to whatever the
+                             * local character set is */
+                            int i;
+
+                            memset(decoded, 0, cliplen);
+                            for (i = 0; ucs2_string[i] != NULL; i++) {
+                                decoded[i] = ucs2_string[i] >= 128 ?
+                                    (*xl_tx[tcsl])(ucs2_string[i]) : ucs2_string[i];
+                            }
+
+                            rc = CopyToClipboard(decoded, cliplen);
+                        }
+                    } else {
+                        CopyToClipboard("", 1);
+                    }
+
+                    free(encoded);
+                    free(decoded);
+                }
+#ifdef KUI
+#ifdef CK_SHELL_NOTIFY
+                else if (tt_clipboard_write == CLIPBOARD_DENY_NOTIFY) {
+                    KuiShowNotification(
+                            KUI_NOTIF_I_WARN,
+                            "Clipboard Write",
+                            "Clipboard write by remote host denied. You can "
+                            "enable clipboard access (or disable this "
+                            "notification) with the SET TERM CLIPBOARD-ACCESS "
+                            "command");
+                }
+#endif /* CK_SHELL_NOTIFY */
+#endif /* KUI */
+            }
+        }
+        break;
+    }
 	case 60: /* XTQALLOWED - query allowed features */
 	case 61: /* XTQDISALLOWED - query disallowed features */
         break;